requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-chess==1.999
openai==1.3.5
requests==2.31.0
python-dotenv==1.0.0
eventlet==0.33.3